refactor(tests): move helper function tests to more appropriate file
