+++ tests/pytest/enrollment/test_views.py
@@ -1,10 +1,8 @@
-from datetime import timedelta
 import time
 
 import pytest
 from authlib.integrations.base_client.errors import UnsupportedTokenTypeError
 from django.urls import reverse
-from django.utils import timezone
 
 from littlepay.api.funding_sources import FundingSourceResponse
 from littlepay.api.groups import GroupFundingSourceResponse
@@ -15,12 +13,6 @@ import benefits.enrollment.views
 import benefits.enrollment.enrollment
 from benefits.core.middleware import TEMPLATE_USER_ERROR
 from benefits.enrollment.views import TEMPLATE_SYSTEM_ERROR, TEMPLATE_RETRY
-from benefits.enrollment.enrollment import (
-    _get_group_funding_source,
-    _calculate_expiry,
-    _is_expired,
-    _is_within_reenrollment_window,
-)
 
 
 @pytest.fixture
@@ -344,30 +336,6 @@ def test_index_eligible_post_valid_form_failure(mocker, client, card_tokenize_fo
 
 
 @pytest.mark.django_db
-@pytest.mark.usefixtures("model_EnrollmentFlow")
-def test_get_group_funding_sources_funding_source_not_enrolled_yet(mocker, mocked_funding_source):
-    mock_client = mocker.Mock()
-    mock_client.get_concession_group_linked_funding_sources.return_value = []
-
-    matching_group_funding_source = _get_group_funding_source(mock_client, "group123", mocked_funding_source.id)
-
-    assert matching_group_funding_source is None
-
-
-@pytest.mark.django_db
-@pytest.mark.usefixtures("model_EnrollmentFlow")
-def test_get_group_funding_sources_funding_source_already_enrolled(
-    mocker, mocked_funding_source, mocked_group_funding_source_no_expiry
-):
-    mock_client = mocker.Mock()
-    mock_client.get_concession_group_linked_funding_sources.return_value = [mocked_group_funding_source_no_expiry]
-
-    matching_group_funding_source = _get_group_funding_source(mock_client, "group123", mocked_funding_source.id)
-
-    assert matching_group_funding_source == mocked_group_funding_source_no_expiry
-
-
-@pytest.mark.django_db
 @pytest.mark.usefixtures("mocked_session_agency", "mocked_session_flow", "mocked_session_eligible")
 def test_index_eligible_post_valid_form_success_does_not_support_expiration_customer_already_enrolled_no_expiry(
     mocker,
@@ -424,32 +392,6 @@ def test_index_eligible_post_valid_form_success_does_not_support_expiration_no_e
     assert (
         model_EnrollmentFlow_does_not_support_expiration.group_id
         in mocked_analytics_module_views.returned_success.call_args.args
-    )
-
-
-def test_calculate_expiry():
-    expiration_days = 365
-
-    expiry_date = _calculate_expiry(expiration_days)
-
-    assert expiry_date == (
-        timezone.localtime(timezone=timezone.get_default_timezone()) + timedelta(days=expiration_days + 1)
-    ).replace(hour=0, minute=0, second=0, microsecond=0)
-
-
-def test_calculate_expiry_specific_date(mocker):
-    expiration_days = 14
-    mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
-        return_value=timezone.make_aware(
-            value=timezone.datetime(2024, 3, 1, 13, 37, 11, 5), timezone=timezone.get_fixed_timezone(offset=0)
-        ),
-    )
-
-    expiry_date = _calculate_expiry(expiration_days)
-
-    assert expiry_date == timezone.make_aware(
-        value=timezone.datetime(2024, 3, 16, 0, 0, 0, 0), timezone=timezone.get_default_timezone()
     )
 
 
@@ -516,42 +458,6 @@ def test_index_eligible_post_valid_form_success_supports_expiration_no_expiry(
     assert model_EnrollmentFlow_supports_expiration.group_id in mocked_analytics_module_views.returned_success.call_args.args
 
 
-def test_is_expired_expiry_date_is_in_the_past(mocker):
-    expiry_date = timezone.make_aware(timezone.datetime(2023, 12, 31), timezone.get_default_timezone())
-
-    # mock datetime of "now" to be specific date for testing
-    mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
-        return_value=timezone.make_aware(timezone.datetime(2024, 1, 1, 10, 30), timezone.get_default_timezone()),
-    )
-
-    assert _is_expired(expiry_date)
-
-
-def test_is_expired_expiry_date_is_in_the_future(mocker):
-    expiry_date = timezone.make_aware(timezone.datetime(2024, 1, 1, 17, 34), timezone.get_default_timezone())
-
-    # mock datetime of "now" to be specific date for testing
-    mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
-        return_value=timezone.make_aware(timezone.datetime(2024, 1, 1, 11, 5), timezone.get_default_timezone()),
-    )
-
-    assert not _is_expired(expiry_date)
-
-
-def test_is_expired_expiry_date_equals_now(mocker):
-    expiry_date = timezone.make_aware(timezone.datetime(2024, 1, 1, 13, 37), timezone.get_default_timezone())
-
-    # mock datetime of "now" to be specific date for testing
-    mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
-        return_value=timezone.make_aware(timezone.datetime(2024, 1, 1, 13, 37), timezone.get_default_timezone()),
-    )
-
-    assert _is_expired(expiry_date)
-
-
 @pytest.mark.django_db
 @pytest.mark.usefixtures("mocked_session_agency", "mocked_session_flow", "mocked_session_eligible")
 def test_index_eligible_post_valid_form_success_supports_expiration_is_expired(
@@ -587,81 +493,6 @@ def test_index_eligible_post_valid_form_success_supports_expiration_is_expired(
     assert response.template_name == model_EnrollmentFlow_supports_expiration.enrollment_success_template
     mocked_analytics_module_views.returned_success.assert_called_once()
     assert model_EnrollmentFlow_supports_expiration.group_id in mocked_analytics_module_views.returned_success.call_args.args
-
-
-def test_is_within_enrollment_window_True(mocker):
-    enrollment_reenrollment_date = timezone.make_aware(timezone.datetime(2023, 2, 1), timezone=timezone.get_default_timezone())
-    expiry_date = timezone.make_aware(timezone.datetime(2023, 3, 1), timezone=timezone.get_default_timezone())
-
-    # mock datetime of "now" to be specific date for testing
-    mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
-        return_value=timezone.make_aware(timezone.datetime(2023, 2, 15, 15, 30), timezone=timezone.get_default_timezone()),
-    )
-
-    is_within_reenrollment_window = _is_within_reenrollment_window(expiry_date, enrollment_reenrollment_date)
-
-    assert is_within_reenrollment_window
-
-
-def test_is_within_enrollment_window_before_window(mocker):
-    enrollment_reenrollment_date = timezone.make_aware(timezone.datetime(2023, 2, 1), timezone=timezone.get_default_timezone())
-    expiry_date = timezone.make_aware(timezone.datetime(2023, 3, 1), timezone=timezone.get_default_timezone())
-
-    # mock datetime of "now" to be specific date for testing
-    mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
-        return_value=timezone.make_aware(timezone.datetime(2023, 1, 15, 15, 30), timezone=timezone.get_default_timezone()),
-    )
-
-    is_within_reenrollment_window = _is_within_reenrollment_window(expiry_date, enrollment_reenrollment_date)
-
-    assert not is_within_reenrollment_window
-
-
-def test_is_within_enrollment_window_after_window(mocker):
-    enrollment_reenrollment_date = timezone.make_aware(timezone.datetime(2023, 2, 1), timezone=timezone.get_default_timezone())
-    expiry_date = timezone.make_aware(timezone.datetime(2023, 3, 1), timezone=timezone.get_default_timezone())
-
-    # mock datetime of "now" to be specific date for testing
-    mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
-        return_value=timezone.make_aware(timezone.datetime(2023, 3, 15, 15, 30), timezone=timezone.get_default_timezone()),
-    )
-
-    is_within_reenrollment_window = _is_within_reenrollment_window(expiry_date, enrollment_reenrollment_date)
-
-    assert not is_within_reenrollment_window
-
-
-def test_is_within_enrollment_window_equal_reenrollment_date(mocker):
-    enrollment_reenrollment_date = timezone.make_aware(timezone.datetime(2023, 2, 1), timezone=timezone.get_default_timezone())
-    expiry_date = timezone.make_aware(timezone.datetime(2023, 3, 1), timezone=timezone.get_default_timezone())
-
-    # mock datetime of "now" to be specific date for testing
-    mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
-        return_value=enrollment_reenrollment_date,
-    )
-
-    is_within_reenrollment_window = _is_within_reenrollment_window(expiry_date, enrollment_reenrollment_date)
-
-    assert is_within_reenrollment_window
-
-
-def test_is_within_enrollment_window_equal_expiry_date(mocker):
-    enrollment_reenrollment_date = timezone.make_aware(timezone.datetime(2023, 2, 1), timezone=timezone.get_default_timezone())
-    expiry_date = timezone.make_aware(timezone.datetime(2023, 3, 1), timezone=timezone.get_default_timezone())
-
-    # mock datetime of "now" to be specific date for testing
-    mocker.patch(
-        "benefits.enrollment.enrollment.timezone.now",
-        return_value=expiry_date,
-    )
-
-    is_within_reenrollment_window = _is_within_reenrollment_window(expiry_date, enrollment_reenrollment_date)
-
-    assert not is_within_reenrollment_window
 
 
 @pytest.mark.django_db
